machine-learning-client/requirements.txt
ibm-watson
pymongo
python-dotenv
